+++ crates/bins/src/bin/datadog-static-analyzer-test-ruleset.rs
@@ -28,6 +28,7 @@ fn test_rule(runtime: &mut JsRuntime, rule: &Rule, test: &RuleTest) -> Result<St
         use_debug: true,
         ignore_generated_files: false,
         timeout: None,
+        debug_java_dfa: false,
     };
     let rules = vec![rule_internal];
     let analyze_result = analyze_with(
