+++ crates/cli/src/file_utils.rs
@@ -594,6 +594,7 @@ mod tests {
             secrets_enabled: false,
             static_analysis_enabled: true,
             secrets_rules: vec![],
+            should_verify_checksum: true,
         };
         assert_eq!(0, filter_files_by_size(&files1, &cli_configuration).len());
